Move all containers into containers folder
