+++ src/containers/App/App.test.js
