+++ src/containers/SearchBar/SearchBar.scss
