+++ src/containers/SearchBar/SearchBar.test.js
